Moved and Added Files
Reason:  The amount of effort needed to wrap everything correctly does
not make it worth it to write in C++.  We would be essentially doing
double coding just in order to do the wrapping.

Moved old cpp files to a new folder.
Added java version of the background files:  ChatRoom, Message, and
ChatRoomManager.