Updated ChatRoomManager and ChatRoomView
The WifiBroadcastReceiver object used by P2PManager needs to be
registered and unregistered when the app closes and resumes. Only the
activity (ChatRoomView) knows when this happens, and needs to make a
call to register and unregister the receiver. Since the activity cannot
directly interface with the P2PManager, it has to go through the
ChatRoomManager which then communicates with the P2PManager. This also
gives functionality to ChatRoomManager to do any cleanup on pausing the
app and any reinitialization on resume.