Final Timestamp Modification
So, I fixed the issue with the display.  It was automatically putting an
image in the display even when there wasn't one.  Since it's now fixed,
my playing with the timestamp is done, but I will not pull until after
our presentation just in case there is a serious issue I might have
overlooked.  I don't want to cause a random failure on this
insignificant change.